Combine GITEA_PUBLIC_URL + GITEA_API_URL -> GITEA_ROOT_URL
This is easier to get right, since it mirrors Gitea's app.ini.

Gitea also has a STATIC_URL_PREFIX which is similar to what GITEA_PUBLIC_URL
was; but we have little motivation to use it, since we are designing
that bids-hook runs _on the same server_ as gitea. Also, if we did keep
it, I would want to more closely mirror app.ini's config behaviour.